Fixes for:
4552 : Trying to Attach a .docx (Word 2007) File Produces: ERROR: value too long for type character varying(32)
4368 : RowId should be marked hidden in Assay List query customize view
4248 : Better Page Names for Editing Pages for Dataset Properties

Includes db update scripts for Core, as the basic problem in 4552 was that the document type column was defined as varchar(32), which is too small for the new Office 2007 mime types (which are 70+ characters). The scripts alter the Documents.DocumentType column. 
 
Also, new icons for the Office 2007 file extensions (e.g., docx, docm, xlsm, xlsb, etc.)