Some tweaks to DataRegionSelection:
We now add a hidden form field with the DataRegion's selection key to grids with record selectors, which means that you don't need to know the query, table, view, and data region name when handling the post.

Hooked things up to use selection state for GridViews that aren't QueryViews.

Made the MS2 controller use the new improved selection state for comparing and exporting runs.