5388: IllegalArgumentException in SamplesController.autoComplete() 5466: Specimen import performance degrades over time on postgres 5184: Specimen reports: new specimen request not appearing in participant report 5447: specimen reports: exception filtering on vial availability 5469: Use "field" terminology in the assay designer 5476: Add "viewName" to the appropriate Client APIs 5486: gridPannelConfig parameter not working in GridView.js
All of the above were code reviewed by Jess.  Note that the changelist makes this look worse than it is; nearly all files are text changes from "property" to "field", and corrections of a public constant called "Particpant" that was used in many files.

Also fixed 5496: FlowManager.vacuum and FlowManager.analyze NPE on Postgres 8.1
This fix wasn't code reviewed, but it's a trivial null check that's blocking my checkin (as well as that of anyone on Postgres 8.1 or 8.3).