Convert CoreController to Spring.  Stop running beehive build step on Pipeline module.
Introduce Project.java, a simple wrapper class for Container that ensures the container is a top-level project (not the root, a subfolder, or null).  We'll eventually use this instead of Container in all places where we actually expect a project.

Clean up some login-related URL handling.

Wiki help views: remove radeox help hack in WikiService, convert html help to jsp.

5573 : SecurityManager.getTermsOfUseHtml() is broken for non-projects
5660 : NPE when running the wiki link checker