5683 : GridView should extend Ext.grid.EditorGridPanel
These new js files define true Ext extensions to their Store and EditorGridPanel classes instead of the wrapped-style used by the existing GridView. Thus users can now call all base-class methods and subscribe to all the base-class events, and the grid can participate in larger Ext layouts. Still need to decide if we want to replace the existing GridView entirely or just depricate it in favor of the new extension.