6186 : NAb details view displays lookup key values in run properties section
This a bigger change than I would like at this stage, but SCHARP has said that the fix is required before the 8.3 release.  The primary changes are isolated to the rendering of the NAb details page.  A minor change was also made to ExprColumn that only has an impact if a PropertyDescriptor name is given a name that is not a valid identifier.

Triage approved and code reviewed by Josh.