for your consideration, since we don't seem to have font size between "smallest on the internet" (aka small) and old man VGA (aka medium).
I played with the themes a little so that MEDIUM is 13px font (seems to be pretty standard) and SMALL and LARGE are smaller increments.

XSMALL is almost identical to old small
SMALL is one step smaller than MEDIUM
MEDIUM is almost identical to old theme
LARGE is much smaller than the previous clown font
XLARGE is also smaller than the previous LARGE, but still big enough to not make the dropdown.

the old themes are still in the dropdown, and the whole thing may be easily reverted.