Issue 11511: Wiki doesn't include attachment deletion in history list Issue 11623: Stop auditing wiki attachment downloads for the guest user
Made auditing of attachments actions more explicitly optional.  Pass in current user if action is to be audited, pass in the null user it is not.    Got rid of a bunch of calls to ..getContext.getUser() in favor of explicit user references.    Cleaned up some old junit tests that were leaving temp files in the build directory.  We also now audit changes to site-wide look-and-feel attachments, so no more posting naughty site logos with impunity.  You have been warned.