12169: SerializationException from org.labkey.api.action.GWTServiceAction.handleRequestInternal()
Small feature to allow users to prevent the server from cleaning up the generated data files that are handed to validation and transform scripts. The files can be useful during development/debugging of a script and there is currently no way to get at them short of setting a breakpoint. There's a new checkbox in the assay designer to control this.