Add popup tooltip that shows reason for conditional formatting if it has been applied.
If a site admin is viewing the pipeline job status grid at the root container and logs out in a separate tab, don't allow the grid to continue refreshing and show the subset that is visible to guests. Guests can't load the page in the first place, so refresh should be consistent.

Delete dead Luminex code.

Issue 10765: **Need documentation annotations for new items in tableInfo.XSD
Issue 12186: **Limit notify list to group