Fixed AncillaryStudyTest and MessagesLongTest failures.
Fixed issue with manageStudyProperties form setting incorrect value for description render type resulting in an exception when trying to load Study Overview web part.

validateForm in CreateAncillaryStudyAction no longer throws a runTimeException, instead it throws an IllegalArgumentException, this way it does not get logged to mothership or show up in the console.