Remove all support for SQL Server 2005 and 2008 (we still support 2008 R2) and Tomcat 5.5.  The server will fail to start (and fail to connect in the case of an external datasource) if these are detected.
Deprecate support for Java 6 and PostgreSQL 8.3. If detected, administrators will see messages in the log and warnings in the header of every page stating they are no longer supported... 13.1 will continue to run correctly. We will remove all support for these in 13.2.

Also, wire up support for using native PostgreSQL functions to implement GROUP_CONCAT... but disable it for now until we resolve an issue with DOMAIN types.