Spec ID: 23109: Share Argos grids and filters with Argos users
Issue 24103: getPolicy on a saved filter entity throws "The requested resource does not exist within this container!" exception

Add a (temporary) workaround to LABKEY.Security.getPolicy() to address the fact that we don't have a general way to resolve a ResourceId to a SecurableResource (only a hack that works for some SecurableResources). Now, if getPolicy() is passed findNearest = false (flag that Susan added recently) and we can't resolve the Resource, then we retrieve the policy directly from the cache and return the assignments (though not the applicable roles).

We need a general SecurableResource resolving mechanism to make this work correctly, but the workaround is good enough for now.