Issue 28694: NPE generates from UNIONized Flow Query
Query doesn't copy the underlying column's FK to a union query column.
Teach GraphColumn to render without relying on a foreign key that describes the graph spec.
As a side effect of this change, graph columns will no longer have a null value if there is no graph for an FCSAnalysis well.
To filter for blank graphs, use a Does Not Start With "~~~" filter.

CR: matt