Issue 28931: Move Tika parsers to a separate classloader Issue 28932: Upgrade Tika
Move tika-app (includes all the parser jars, the vast majority of the Tika bulk) to the Search module. Leave behind just tika-core in API. Use the recently added separate classloader in Search to keep all the parser jars off the main classpath.

Upgrade Tika 1.5 -> 1.14

Upgrade ROME (1.0 -> 1.7.1) and add JDOM 2.0.4 dependency

Add jars for direct dependencies for which we used to rely on tika-app-1.5. Upgrade them to the latest versions:
- Apache Commons Compress 1.13 (upgrade from 1.5)
- Apache PDFBox 2.0.4 (upgrade from 1.8.4)
- Apache POI 3.15 (upgrade from 3.10-beta2)